backend/pyproject.toml
[project]
name = "openkms-backend"
version = "0.1.0"
description = "openKMS FastAPI backend - document management, pipelines, jobs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "boto3>=1.35.0",
    "botocore>=1.35.0",
    "python-dotenv>=1.0.0",
    "itsdangerous>=2.0.0",
    "procrastinate>=2.0.0",
    "psycopg[binary]>=3.0.0",
    "openai>=1.0.0",
    "pydantic-ai>=0.0.14",
    "PyJWT[crypto]>=2.8.0",
    "pgvector>=0.4.2",
    "cryptography>=44.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
